timestep debug

+++ main/src/propagator/std_disk.hpp
@@ -166,9 +166,9 @@ public:
         timer.step("duTimestep");
 
         double t_du_all;
-        double t_starAll;
+        float t_starAll;
         MPI_Allreduce(&star.t_du, &t_du_all, 1, MpiType<double>{}, MPI_MIN, MPI_COMM_WORLD);
-        MPI_Allreduce(&star.t_star, &t_starAll, 1, MpiType<double>{}, MPI_MIN, MPI_COMM_WORLD);
+        MPI_Allreduce(&star.t_star, &t_starAll, 1, MpiType<float>{}, MPI_MIN, MPI_COMM_WORLD);
 
         if (Base::rank_ == 0) {
             printf("t_du_all: %lf\n, t_starAll: %lf\n", t_du_all, t_starAll);
